+++ plutarch-extra/Plutarch/Extra/Map.hs
@@ -1,5 +1,4 @@
 {-# LANGUAGE QuantifiedConstraints #-}
-{-# OPTIONS_GHC -Wno-orphans #-}
 
 module Plutarch.Extra.Map (
   -- * Lookup
@@ -19,8 +18,6 @@ module Plutarch.Extra.Map (
   pfoldlWithKey,
 
   -- * Conversion
-  punsortedMapFromFoldable,
-  psortedMapFromFoldable,
   pkeys,
 
   -- * Key-value pair manipulation
@@ -29,13 +26,9 @@ module Plutarch.Extra.Map (
   pkvPairLt,
 ) where
 
-import Data.Foldable (foldl')
-import GHC.Exts (IsList (Item, fromList, toList))
 import Plutarch.Api.V1.AssocMap (
   KeyGuarantees (Sorted, Unsorted),
   PMap (PMap),
-  pempty,
-  pinsert,
   plookup,
  )
 import Plutarch.Builtin (ppairDataBuiltin)
@@ -207,62 +200,6 @@ ptryLookup = phoistAcyclic $
   plam $ \k kvs ->
     passertPJust # "plookupPartial: No value found for key."
       # (plookup # k # kvs)
-
-{- | Given a 'Foldable' of key-value pairs, construct an unsorted 'PMap'.
- Performs linearly with respect to its argument.
-
- = Note
-
- If there are duplicate keys in the input, the /last/ key will \'win\' in a
- lookup.
--}
-punsortedMapFromFoldable ::
-  forall (k :: PType) (v :: PType) (f :: Type -> Type) (s :: S).
-  (Foldable f, PIsData k, PIsData v) =>
-  f (Term s k, Term s v) ->
-  Term s (PMap 'Unsorted k v)
-punsortedMapFromFoldable = pcon . PMap . foldl' go (pcon PNil)
-  where
-    go ::
-      forall (s' :: S).
-      Term s' (PBuiltinList (PBuiltinPair (PAsData k) (PAsData v))) ->
-      (Term s' k, Term s' v) ->
-      Term s' (PBuiltinList (PBuiltinPair (PAsData k) (PAsData v)))
-    go acc (key, val) =
-      pcon . PCons (ppairDataBuiltin # pdata key # pdata val) $ acc
-
-{- | Given a 'Foldable' of (not necessarily sorted) key-value pairs, construct a
- 'PMap' which is guaranteed sorted. Performs a linear number of ordered
- insertions with respect to the length of its argument.
-
- = Note
-
- If there are duplicate keys, only the /last/ key-value pair will remain in
- the result.
--}
-psortedMapFromFoldable ::
-  forall (k :: PType) (v :: PType) (f :: Type -> Type) (s :: S).
-  (Foldable f, POrd k, PIsData k, PIsData v) =>
-  f (Term s k, Term s v) ->
-  Term s (PMap 'Sorted k v)
-psortedMapFromFoldable = foldl' go pempty
-  where
-    go ::
-      forall (s' :: S).
-      Term s' (PMap 'Sorted k v) ->
-      (Term s' k, Term s' v) ->
-      Term s' (PMap 'Sorted k v)
-    go acc (key, val) = pinsert # key # val # acc
-
-instance (PIsData k, PIsData v, POrd k) => IsList (Term s (PMap 'Unsorted k v)) where
-  type Item (Term s (PMap 'Unsorted k v)) = (Term s k, Term s v)
-  fromList = punsortedMapFromFoldable
-  toList = error "unimplemented"
-
-instance (PIsData k, PIsData v, POrd k) => IsList (Term s (PMap 'Sorted k v)) where
-  type Item (Term s (PMap 'Sorted k v)) = (Term s k, Term s v)
-  fromList = psortedMapFromFoldable
-  toList = error "unimplemented"
 
 {- | Get a list-like structure full of the keys of the argument 'PMap'. If the
  'PMap' is 'Sorted', the keys will maintain that order, and will be unique;
